# perf: Micro optimization EndsWith using