--- a/src/Avalonia.Build.Tasks/Avalonia.Build.Tasks.csproj
+++ b/src/Avalonia.Build.Tasks/Avalonia.Build.Tasks.csproj
@@ -125,9 +125,5 @@
       <PackageReference Include="Microsoft.Build.Framework" Version="15.1.548" PrivateAssets="All" />
       <PackageReference Include="System.Numerics.Vectors" Version="4.5.0" />
     </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="Compatibility\" />
-    </ItemGroup>
   <Import Project="..\..\build\SourceGenerators.props" />
 </Project>