<!--- See CONTRIBUTING.md for general guidelines on contributions -->

## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->

- Replace `ToLowerInvariant().EndsWith` with `EndsWith("string", StringComparison.OrdinalIgnoreCase)`
- Replace `value.EndsWith("%")` with `value[value.Length -1 ] == '%'`

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->


## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->

small decrease in memory usage during build

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->

```

BenchmarkDotNet v0.13.8, Windows 10 (10.0.19045.3448/22H2/2022Update)
Intel Core i5-10400 CPU 2.90GHz, 1 CPU, 12 logical and 6 physical cores
.NET SDK 7.0.401
  [Host]     : .NET 6.0.22 (6.0.2223.42425), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.22 (6.0.2223.42425), X64 RyuJIT AVX2


```
| Method                         | Mean       | Error     | StdDev    | Median     | Ratio | Gen0      | Allocated  | Alloc Ratio |
|------------------------------- |-----------:|----------:|----------:|-----------:|------:|----------:|-----------:|------------:|
| ToLowerInvariant_EndsWith      | 193.712 ms | 3.3250 ms | 3.1102 ms | 193.146 ms |  1.00 | 6000.0000 | 38600893 B |       1.000 |
| EndsWith_With_StringComparison |   3.959 ms | 0.2222 ms | 0.6411 ms |   3.761 ms |  0.02 |         - |        2 B |       0.000 |

```

BenchmarkDotNet v0.13.8, Windows 10 (10.0.19045.3448/22H2/2022Update)
Intel Core i5-10400 CPU 2.90GHz, 1 CPU, 12 logical and 6 physical cores
.NET SDK 7.0.401
  [Host]     : .NET 6.0.22 (6.0.2223.42425), X64 RyuJIT AVX2
  DefaultJob : .NET 6.0.22 (6.0.2223.42425), X64 RyuJIT AVX2


```
| Method                     | Mean     | Error     | StdDev    | Median   | Ratio | RatioSD | Allocated | Alloc Ratio |
|--------------------------- |---------:|----------:|----------:|---------:|------:|--------:|----------:|------------:|
| InvariantCultureIgnoreCase | 4.064 ms | 0.2365 ms | 0.6935 ms | 3.874 ms |  1.00 |    0.00 |       2 B |        1.00 |
| OrdinalIgnoreCase          | 1.490 ms | 0.0280 ms | 0.0691 ms | 1.457 ms |  0.36 |    0.06 |       1 B |        0.50 |



## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/avalonia-docs with user documentation

## Breaking changes
<!--- List any breaking changes here. -->

## Obsoletions / Deprecations
<!--- Obsolete and Deprecated attributes on APIs MUST only be included when discussed with Core team. @grokys, @kekekeks & @danwalmsley -->

## Fixed issues
<!--- If the pull request fixes issue(s) list them like this: 
Fixes #123
Fixes #456
-->