@avaloniaui-team 2023-09-30
You can test this PR using the following package version. `11.0.999-cibuild0040249-beta`. (feed url: https://nuget-feed-all.avaloniaui.net/v3/index.json) [PRBUILDID]