@avaloniaui-team 2023-10-01
You can test this PR using the following package version. `11.0.999-cibuild0040342-beta`. (feed url: https://nuget-feed-all.avaloniaui.net/v3/index.json) [PRBUILDID]